rename: hook directory to hooks
